Split the `onnx` package into an API and an implementation under `internal/`, so it can be further split.
