+++ internal/onnxgomlx/external_data_reader.go
@@ -1,4 +1,4 @@
-package onnx
+package onnxgomlx
 
 import (
 	"io"
