+++ internal/onnxgomlx/fusion_dense_gelu.go
@@ -1,4 +1,4 @@
-package onnx
+package onnxgomlx
 
 import (
 	. "github.com/gomlx/gomlx/pkg/core/graph" //nolint
