+++ internal/onnxgomlx/fusion_quantized_dense.go
@@ -1,4 +1,4 @@
-package onnx
+package onnxgomlx
 
 import (
 	"math"
