+++ internal/onnxgomlx/fusion_quantized_qkv.go
@@ -1,4 +1,4 @@
-package onnx
+package onnxgomlx
 
 import (
 	"github.com/gomlx/gomlx/backends"
