+++ internal/onnxgomlx/fusion_test.go
@@ -1,4 +1,4 @@
-package onnx
+package onnxgomlx
 
 import (
 	"math"
