+++ internal/onnxgomlx/materialize.go
@@ -1,4 +1,4 @@
-package onnx
+package onnxgomlx
 
 import (
 	"fmt"
