+++ internal/onnxgomlx/ops.go
@@ -1,4 +1,4 @@
-package onnx
+package onnxgomlx
 
 import (
 	"fmt"
