+++ internal/onnxgomlx/tensor.go
@@ -1,4 +1,4 @@
-package onnx
+package onnxgomlx
 
 import (
 	"encoding/binary"
