+++ internal/onnxgomlx/variables.go
@@ -1,21 +1,19 @@
-package onnx
+package onnxgomlx
 
 import (
 	"strings"
 
 	"github.com/gomlx/exceptions"
 	"github.com/gomlx/gomlx/pkg/ml/context"
+	"github.com/gomlx/onnx-gomlx/onnx"
 	"github.com/pkg/errors"
 )
 
 // This file defines importing variables from ONNX and (TODO) saving them back to the ONNX model file.
 
-// ModelScope is the default model scope to use when for the ONNX model variables when converting to GoMLX.
-var ModelScope = "ONNX"
-
 // This file defines the methods that build the computation graph using GoMLX.
 
-// VariablesToContext will create variables in the context (within scope ModelScope) from
+// VariablesToContext will create variables in the context (within scope onnx.ModelScope) from
 // all variables present in the model initializer list.
 //
 // Call this once in your context, before using the model with Model.CallGraph.
@@ -24,9 +22,9 @@ var ModelScope = "ONNX"
 // See also ContextToONNX, if after converting and fine-tuning an ONNX model, you want to update its weights.
 func (m *Model) VariablesToContext(ctx *context.Context) error {
 	if len(m.Proto.Graph.SparseInitializer) > 0 {
-		exceptions.Panicf("onnx.VariablesToContext does not support ONNX SparseTensors")
+		exceptions.Panicf("onnxgomlx.VariablesToContext does not support ONNX SparseTensors")
 	}
-	ctx = ctx.In(ModelScope).Checked(false)
+	ctx = ctx.In(onnx.ModelScope).Checked(false)
 	reader := m.getExternalDataReader()
 	for _, tensorProto := range m.Proto.Graph.Initializer {
 		tensor, err := tensorToGoMLXWithBaseDir(m.backend, tensorProto, m.baseDir(), reader)
@@ -87,9 +85,9 @@ func (m *Model) FreeUnusedVariables() {
 // moving averages) are converted.
 func (m *Model) ContextToONNX(ctx *context.Context) error {
 	if len(m.Proto.Graph.SparseInitializer) > 0 {
-		exceptions.Panicf("onnx.VariablesToContext does not support ONNX SparseTensors")
+		exceptions.Panicf("onnxgomlx.VariablesToContext does not support ONNX SparseTensors")
 	}
-	ctx = ctx.In(ModelScope)
+	ctx = ctx.In(onnx.ModelScope)
 	for _, tensorProto := range m.Proto.Graph.Initializer {
 		tensorName := SafeVarName(tensorProto.Name)
 		gomlxVar := ctx.GetVariable(tensorName)
